fix so simpler and works using %run in Jupyter
Makes so can use %run in Jupyter and still works with %%bash magic in Jupyter and also in the terminal. It seems the `Live()` option didn't work well from using `%run` in Jupyter or you get duplicate displays in if you use printing of the terminal along with print. This seems more consistent and still only requites installing Rich if want to use display.