add mono setting for display without color
The version without any font color looks so nice I figured might as well make a way for user's to opt in when using it in the equivalent of the terminal. I figure some users may not like the colors applied by rich-dataframe and find the simpler, cleaner look may work without them needing to adjust anything.